Move everything done so far into a game/ folder
