+++ CMakeLists.txt
@@ -4,7 +4,12 @@ set(CMAKE_CXX_STANDARD 17)
 
 project(graphisoft_racer)
 
-add_executable(graphisoft_racer src/car.cpp src/driving.cpp src/map.cpp src/vector.cpp src/main.cpp)
+add_executable(graphisoft_racer
+    src/game/car.cpp
+    src/game/driving.cpp
+    src/game/map.cpp
+    src/game/vector.cpp
+    src/main.cpp)
 
 # Ez miert nem mukodik? Mindegy, majd Adam megszereli
 set_target_properties(graphisoft_racer PROPERTIES VS_DEBUGGER_WORKING_DIRECTORY "${CMAKE_SOURCE_DIR}")
