+++ src/game/vector.cpp
