+++ src/game/vector.h
